The portable core runs on the second vendor's silicon
NUCLEO-F446RE flashed and answering. First real number off the STM32:
vdda = 3303.8 mV, computed from the VREFINT calibration ST burns in at
0x1FFF7A2A rather than assuming 3300. Two flashes read 3303.7 and 3303.8,
so the path repeats to 0.1 mV. Not yet cross-checked against a multimeter;
that is the next step of day 12.

Three days of "the computer does not see the board" were a cable. It held
VBUS — LD3 was lit the whole time — and lost data under load, so the
ST-LINK re-enumerated roughly once a second, dying during driver matching:
right after SET_CONFIGURATION, the moment it starts drawing full current.
Thirty consecutive ioreg samples each returned a *new* registry node id.
That is what ruled out macOS permissions, which break open(), not presence
on the bus. ST-LINK firmware went V2J33M25 (2019) to V2J48M35 on the way.

Flashed by copying the .bin onto the NOD_F446RE mass-storage volume. The
CubeIDE debug session still refuses to start without ST-Link Server, and
the stlinkserver plugin for macOS/aarch64 ships a 3.5 KB resolver class
with no binary behind it. Programming does not need it; breakpoints do.

Running on hardware immediately found a bug the host build cannot:
PRIu64 in `status` expands to %llu, which newlib-nano does not support
without -u _printf_long_long. It does not merely skip the number — it
never consumes the argument, so everything after it shifts one position.
The line still looked plausible:

  OK,status,state=IDLE,uptime_s=lu,accepted=39,failed=0,gaps=0

where 39 was the uptime, `failed` held accepted, and the real gaps value
was never printed at all. Narrowed to uint32_t; seconds overflow it after
136 years, so linking the 64-bit formatter would buy nothing.

Third instance in this project of a tool that does not fail, but quietly
does the wrong thing: CDT registering a source folder and generating no
rules for it, newlib dropping %f, newlib dropping %llu.

+++ firmware/app/stm32/src/app_stm32.c
@@ -207,9 +207,17 @@ static void cmd_status(cli_t *cli, int argc, char **argv)
     (void)argc;
     (void)argv;
     snprintf(buf, sizeof buf,
-             "OK,status,state=%s,uptime_s=%" PRIu64 ",accepted=%" PRIu32
+             "OK,status,state=%s,uptime_s=%" PRIu32 ",accepted=%" PRIu32
              ",failed=%" PRIu32 ",gaps=%" PRIu32 "\n",
-             session_state_name(&s_session), platform_time_us() / 1000000u,
+             session_state_name(&s_session),
+             /*
+              * Deliberately narrowed to 32 bits. newlib-nano drops %llu
+              * unless the image is linked with -u _printf_long_long, and it
+              * drops it silently: the board printed "uptime_s=lu" for a week
+              * before anyone ran it on hardware. Seconds overflow uint32_t
+              * after 136 years, so the 64-bit formatter buys nothing here.
+              */
+             (uint32_t)(platform_time_us() / 1000000u),
              s_sampler.accepted, s_sampler.sensor_failures, s_sampler.gaps);
     cli_write(cli, buf);
 }
